feat(tagging): finalize most of the backend for now (#13)
Reviewed-on: https://gitea.ahmadfarhan.com/paan/LinkStash/pulls/13

+++ linkstash_backend/src/controllers/bookmark.controller.ts
@@ -1,18 +1,19 @@
 //#region Imports
 import {authenticate} from '@loopback/authentication';
 import {inject} from '@loopback/core';
-import {Count, CountSchema, Filter, FilterBuilder, FilterExcludingWhere, Where, repository} from '@loopback/repository';
-import {del, get, getModelSchemaRef, param, patch, post, requestBody, response} from '@loopback/rest';
+import {Filter, FilterBuilder, FilterExcludingWhere, repository} from '@loopback/repository';
+import {Response, RestBindings, del, get, getModelSchemaRef, param, patch, post, requestBody, response} from '@loopback/rest';
 import {SecurityBindings, UserProfile, securityId} from '@loopback/security';
 import {Bookmark, BookmarkWithRelations} from '../models';
-import {BookmarkRepository} from '../repositories';
+import {BookmarkRepository, UserRepository} from '../repositories';
+import {bookmarkPatchSchema} from '../types';
 //#endregion
 
 @authenticate('jwt')
 export class BookmarkController {
   constructor(
-    @repository(BookmarkRepository)
-    public bookmarkRepository: BookmarkRepository,
+    @repository(BookmarkRepository) public bookmarkRepository: BookmarkRepository,
+    @repository(UserRepository) public userRepository: UserRepository,
   ) {}
 
   //#region CRUD related endpoints
@@ -37,6 +38,7 @@ export class BookmarkController {
   ): Promise<Bookmark> {
     bookmark.userId = currentUserProfile[securityId];
     return this.bookmarkRepository.create(bookmark);
+    //return this.userRepository.bookmarks(id).create(bookmark);
   }
 
   // @intercept('interceptors.AddCountToResultInterceptor')
@@ -53,17 +55,13 @@ export class BookmarkController {
     },
   })
   async find(@inject(SecurityBindings.USER) currentUserProfile: UserProfile, @param.filter(Bookmark) filter?: Filter<Bookmark>): Promise<Object> {
-    const builder = new FilterBuilder(filter);
-    builder.impose({userId: currentUserProfile[securityId]});
+    const builder = new FilterBuilder(filter).impose({userId: currentUserProfile[securityId]});
     const resultFilter = builder.build();
     const allResultFilter: Filter<Bookmark> = JSON.parse(JSON.stringify(resultFilter));
     if (allResultFilter.limit) delete allResultFilter.limit;
 
-    console.log(JSON.stringify(allResultFilter));
     const all = await this.bookmarkRepository.find(allResultFilter);
-
     const returnValue = {
-      ...filter,
       countAll: all.length,
       data: all,
     };
@@ -81,10 +79,17 @@ export class BookmarkController {
   })
   async findById(
     @param.path.number('id') id: number,
-    @param.filter(Bookmark, {exclude: 'where'})
-    filter?: FilterExcludingWhere<Bookmark>,
+    @inject(SecurityBindings.USER) currentUserProfile: UserProfile,
+    @repository(UserRepository) userRepository: UserRepository,
+    @inject(RestBindings.Http.RESPONSE) res: Response,
+    @param.filter(Bookmark, {exclude: 'where'}) filter?: FilterExcludingWhere<Bookmark>,
   ): Promise<BookmarkWithRelations> {
-    return this.bookmarkRepository.findById(id, filter);
+    const combinedUserFilter = new FilterBuilder<Bookmark>(filter).impose({userId: currentUserProfile[securityId], id: id}).build();
+    const results = await userRepository.bookmarks(currentUserProfile['securityId']).find(combinedUserFilter);
+    if (results.length > 0) return results[0];
+
+    res.status(404);
+    throw new Error(`Entity not found: Bookmark with id ${id}`);
   }
 
   // @intercept('interceptors.AddCountToResultInterceptor')
@@ -121,74 +126,52 @@ export class BookmarkController {
     @requestBody({
       content: {
         'application/json': {
-          schema: getModelSchemaRef(Bookmark, {partial: true}),
+          schema: bookmarkPatchSchema,
         },
       },
     })
     bookmark: Bookmark,
+    @inject(SecurityBindings.USER) currentUserProfile: UserProfile,
+    @inject(RestBindings.Http.RESPONSE) res: Response,
   ): Promise<void> {
-    await this.bookmarkRepository.updateById(id, bookmark);
+    // console.log(bookmark);
+    const existing = await this.userRepository.bookmarks(currentUserProfile[securityId]).find({where: {id: id}});
+    // console.log(existing);
+    if (existing.length > 0) {
+      await this.bookmarkRepository.updateAll(bookmark, {id: id});
+    } else {
+      res.status(404);
+      throw new Error(`Entity not found: Bookmark with id ${id}`);
+    }
   }
 
   @del('/bookmarks/{id}')
   @response(204, {
     description: 'Bookmark DELETE success',
   })
-  async deleteById(@param.path.number('id') id: number): Promise<void> {
-    await this.bookmarkRepository.deleteById(id);
+  async deleteById(
+    @param.path.number('id') id: number,
+    @inject(SecurityBindings.USER) currentUserProfile: UserProfile,
+    @inject(RestBindings.Http.RESPONSE) res: Response,
+  ): Promise<void> {
+    const existing = await this.userRepository.bookmarks(currentUserProfile[securityId]).find({where: {id: id}});
+    if (existing.length > 0) {
+      await this.bookmarkRepository.deleteById(id);
+    } else {
+      res.status(404);
+      throw new Error(`Entity not found: Bookmark with id ${id}`);
+    }
+
+    //await this.bookmarkRepository.deleteById(id);
   }
 
-  @get('/bookmarks/count')
-  @response(200, {
-    description: 'Bookmark model count',
-    content: {'application/json': {schema: CountSchema}},
-  })
-  async count(@param.where(Bookmark) where?: Where<Bookmark>): Promise<Count> {
-    return this.bookmarkRepository.count(where);
-  }
-  //#endregion
-
-  //#region Users related endpoints
-  // @get('/bookmarks/{id}/user', {
-  //   responses: {
-  //     '200': {
-  //       description: 'User belonging to Bookmark',
-  //       content: {
-  //         'application/json': {
-  //           schema: getModelSchemaRef(User),
-  //         },
-  //       },
-  //     },
-  //   },
-  // })
-  // async getUser(
-  //   @param.path.number('id') id: typeof Bookmark.prototype.id,
-  // ): Promise<User> {
-  //   return this.bookmarkRepository.user(id);
-  // }
-  //#endregion
-
-  /**
-   * ! Everything below this line are default LB4 implementation. Everything above is customized for LinkStash
-   * TODO verify that it fits our needs
-   */
-
-  // @patch('/bookmarks')
+  // @get('/bookmarks/count')
   // @response(200, {
-  //   description: 'Bookmark PATCH success count',
+  //   description: 'Bookmark model count',
   //   content: {'application/json': {schema: CountSchema}},
   // })
-  // async updateAll(
-  //   @requestBody({
-  //     content: {
-  //       'application/json': {
-  //         schema: getModelSchemaRef(Bookmark, {partial: true}),
-  //       },
-  //     },
-  //   })
-  //   bookmark: Bookmark,
-  //   @param.where(Bookmark) where?: Where<Bookmark>,
-  // ): Promise<Count> {
-  //   return this.bookmarkRepository.updateAll(bookmark, where);
+  // async count(@param.where(Bookmark) where?: Where<Bookmark>): Promise<Count> {
+  //   return this.bookmarkRepository.count(where);
   // }
+  //#endregion
 }
